New completion script for quest Plagued Lands.
The npc does not reset his orientation at the end of the script. That is why he has a different orientation in every sniff. When somebody turns in this quest, the npc turns to face the captured bear, and he remains facing that direction until he respawns. I killed him and waited for him to respawn to get the correct default orientation.